optimised for build template
