+++ generate/test/CMakeLists.txt
@@ -20,6 +20,7 @@ ADD_EXECUTABLE(mnist
 ADD_EXECUTABLE(nn
 				${PROJECT_SOURCE_DIR}/example/nn/index.cpp)
 if(MSVC)
+target_link_libraries(nn cactus)
 else()
 target_link_libraries(cpptest pthread)
 endif()
