+++ src/core/framework/node.hpp
@@ -3,7 +3,7 @@
 #define SRC_NODE_HPP_
 
 #include "tensor.hpp"
-#include <cstring>
+#include <string>
 #include <iostream>
 #include <vector>
 //#include <vector>
@@ -16,24 +16,18 @@ namespace cactus {
     enum ComputeStatus { CtNoCompute, CtComputed, CtQueuing };
     class Node {
     public:
-        Node() : cstatus_(CtNoCompute) {}
-        virtual std::size_t num_inputs() { return 0; }
-        virtual Node *input(std::size_t) { return NULL; }
-        ComputeStatus status() { return cstatus_; };
-        void status(ComputeStatus cs) { cstatus_ = cs; };
+        Node();
+        virtual std::size_t num_inputs();
+        virtual Node *input(std::size_t);
+        ComputeStatus status();
+        void status(ComputeStatus cs);
         virtual NodeType type() = 0;
-        void run() {
-            compute();
-            status(CtComputed);
-        }
-        virtual void compute() { std::cout << "Node:" << type() << std::endl; }
-        virtual xgrads grad(Tensor&) {
-            return xgrads();
-        }
-        Tensor &tensor() { return t; }
-        std::string name() { return name_; }
-        void name(std::string v) { name_ = v; }
-
+        void run();
+        virtual void compute();
+        virtual xgrads grad(Tensor&);
+        Tensor &tensor();
+        std::string name();
+        void name(std::string v);
     protected:
         Tensor t;
         std::string name_;
@@ -42,64 +36,37 @@ namespace cactus {
 
     class NodePlaceholder : public Node {
     public:
-        NodePlaceholder(Shape s, DataType dt) {}
-        NodeType type() { return NtPlaceholder; }
-        void assign(const Tensor& v) {
-            t = v;
-            status(CtComputed);
-        }
+        NodePlaceholder(Shape s, DataType dt);
+        NodeType type();
+        void assign(const Tensor& v);
     };
     class NodeVariable : public Node {
     public:
-        NodeVariable(const Tensor &v) { init_val = v; }
-        NodeType type() { return NtVariable; }
-        virtual void compute() {
-            t = init_val;
-            status(CtComputed);
-        }
+        NodeVariable(const Tensor &v);
+        NodeType type();
+        virtual void compute();
     private:
         Tensor init_val;
     };
     class NodeConst : public Node {
     public:
-        NodeConst(const Tensor &v) { cstatus_ = (CtNoCompute); t = v; }
-        NodeType type() { return NtConst; }
+        NodeConst(const Tensor &v);
+        NodeType type();
     };
-
-    class Output {
+    
+    class Cx_EXPORTS Output {
     public:
-        Output() : node_(NULL) {}
-        Output(const Output &o) { node_ = o.node(); }
-        Output(Node *n) : node_(n) {}
-        Node *node() const {
-            assert(node_ != NULL);
-            return node_;
-        }
-        Tensor& tensor() {
-            assert(node_ != NULL);
-            return node_->tensor();
-        }
-        template<typename T>
-        static void dump(std::ostream &out,Tensor& c) {
-            auto s = c.shape();
-            for (size_t i = 0; i < s.rows; i++)
-            {
-                for (size_t j = 0; j < s.cols; j++)
-                {
-                    out << c.get<T>(i*s.rows+j);
-                }
-            }
-            out << std::endl;
-        }
-        friend std::ostream& operator<<(std::ostream &out, Output &c) {
-            CASES(c.tensor().dtype(), dump<T>(out,c.tensor()));
-            return out;
-        }
+        Output();
+        Output(const Output &o);
+        Output(Node *n);
+        Node *node() const;
+        Tensor& tensor();
+        void dump();
     private:
         Node *node_;
     };
 
-    class Input {
+    class Cx_EXPORTS Input {
     public:
         struct Initializer {
             template <typename T> Initializer(const T &v) {
@@ -112,27 +79,14 @@ namespace cactus {
                 tensor = Tensor(DataTypeToEnum<T>::value, s);
                 std::memcpy(tensor.data(), v.begin(), tensor.totalBytes());
             }
-            Initializer(const Output& v) {
-                tensor.assign(v.node()->tensor());
-            }
-            Initializer(const std::initializer_list<Initializer> &v) {
-                uint32_t offset = 0;
-                auto const &first = *v.begin();
-                Shape s = { v.size(),first.tensor.shape().cols  };
-                tensor = Tensor(first.tensor.dtype(), s);
-
-                for (auto n : v) {
-                    std::memcpy((char *)tensor.data() + offset, n.tensor.data(),
-                        n.tensor.totalBytes());
-                    offset += n.tensor.totalBytes();
-                }
-            }
+            Initializer(const Output& v);
+            Initializer(const std::initializer_list<Initializer> &v);
             Tensor tensor;
         };
 
     public:
-        Input(Output out) : output_(out) {}
-        Node *node() const { return output_.node(); }
+        Input(Output out);
+        Node *node() const;
         // Input(const Initializer& init){}
         /*template<typename T>
         Input(const T& v): Input(Initializer(v)) {}*/
@@ -140,26 +94,21 @@ namespace cactus {
             return tensor_;
         }*/
     private:
-        Input() {}
+        Input();
         Output output_;
         // Tensor tensor_;
     };
 
     class Operation : public Node {
     public:
-        void input(Node *v) { inputs.push_back(v); }
-        Node *input(std::size_t idx) {
-            if (idx < inputs.size()) {
-                return inputs.at(idx);
-            }
-            return NULL;
-        }
-        std::size_t num_inputs() { return inputs.size(); }
+        void input(Node *v);
+        Node *input(std::size_t idx);
+        std::size_t num_inputs();
         /*Output output(uint32_t idx) {
             return Output(this);
         }*/
 
-        NodeType type() { return NtOperation; }
+        NodeType type();
 
     protected:
         std::vector<Node *> inputs;
