--- src/core/framework/session.hpp
@@ -1,22 +1,0 @@
-/*
-Copyright(C) 2018 liuwenjun.All rights reserved.
-*/
-
-#ifndef SRC_SESSION_HPP_
-#define SRC_SESSION_HPP_
-
-#include <string>
-#include <vector>
-#include <utility>
-
-namespace cactus {
-
-class session {
- public:
-    /*int run(const std::vector<std::pair<std::string, tensor> >& inputs) {
-        return 0;
-    }*/
-};
-}
-
-#endif  // SRC_SESSION_HPP_
