--- src/core/kernels/math.cpp
@@ -1,4 +1,0 @@
-#include "math.h"
-
-
-
