--- src/core/kernels/ops_graph.cpp
@@ -1,5 +1,0 @@
-#include "ops_graph.h"
-namespace cactus {
-
-
-} // namespace cactus
