--- src/core/kernels/ops_graph.h
@@ -1,21 +1,0 @@
-#include "../framework/node.hpp"
-
-namespace cactus {
-class InitVariable : public Operation {
-  public:
-      InitVariable(std::vector<Node *> l) { inputs = l; }
-    void compute() {
-        for (auto n : inputs) {
-            switch (n->type()) {
-            case NtVariable:
-                n->compute();
-                break;
-            default:
-                break;
-            }
-        }
-    }
-    NodeType type() { return NtInitOp; }
-};
-
-} // namespace cactus
